+++ libft/ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: seojkim <seojkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 14:44:13 by seojkim           #+#    #+#             */
-/*   Updated: 2023/11/07 13:15:00 by seojkim          ###   ########.fr       */
+/*   Updated: 2023/11/12 17:06:22 by seojkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(unsigned int count, unsigned int size)
+void	*ft_calloc(unsigned int count, size_t size)
 {
 	void	*address;
