+++ libft/ft_memchr.c
@@ -6,13 +6,13 @@
 /*   By: seojkim <seojkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 15:07:41 by seojkim           #+#    #+#             */
-/*   Updated: 2023/11/09 13:15:10 by seojkim          ###   ########.fr       */
+/*   Updated: 2023/11/12 17:06:48 by seojkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*str;
