+++ libft/ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: seojkim <seojkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 15:08:29 by seojkim           #+#    #+#             */
-/*   Updated: 2023/11/06 16:36:10 by seojkim          ###   ########.fr       */
+/*   Updated: 2023/11/12 17:07:03 by seojkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, unsigned int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*ptr;
