+++ libft/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: seojkim <seojkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 17:57:55 by seojkim           #+#    #+#             */
-/*   Updated: 2023/11/06 16:44:37 by seojkim          ###   ########.fr       */
+/*   Updated: 2023/11/12 17:07:24 by seojkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	int	i;
