+++ libft/ft_strnstr.c
@@ -6,13 +6,13 @@
 /*   By: seojkim <seojkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 18:43:52 by seojkim           #+#    #+#             */
-/*   Updated: 2023/11/06 16:36:36 by seojkim          ###   ########.fr       */
+/*   Updated: 2023/11/12 17:08:36 by seojkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
 	unsigned int	n_index;
 	unsigned int	h_index;
